moved Twig trans extension to Symfony\Bridge
