+++ src/Symfony/Bridge/Twig/Node/TransNode.php
@@ -3,18 +3,18 @@
 /*
  * This file is part of the Symfony package.
  *
- * (c) Fabien Potencier <fabien@symfony.com>
+ * (c) Fabien Potencier <fabien.potencier@symfony-project.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\TwigBundle\Node;
+namespace Symfony\Bridge\Twig\Node;
 
 /**
- * 
  *
- * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @author Fabien Potencier <fabien.potencier@symfony-project.com>
  */
 class TransNode extends \Twig_Node
 {
